Do not start in an X11 session and opt out of session managment
xwaylandvideobridge runs with the Qt xcb backend, this means it
is also session managed by default. Since it is also autostarted
on every log a new instance is auotstarted and in addition the
session manager will start an instance for every one that was
running. So on second log in you will have 2 running instances,
on the third three, ...
Also exit on X11 where it is not needed.