Remove the tray icon
This was useful when this was first created, and didn't auto-start. You
would start XWaylandVideoBridge manually, and then had to stop it
manually as well.

But nowadays, this is auto-started and will always run in the background
and has no visible user interface. Additionally, the only thing the tray
icon can do is make it more obvious you can quit the bridge - which you
almost never should do. Then it couldn't do it's job!